backend/requirements.txt
aio-pika==7.2.0
aiomysql==0.0.22
aioredis==2.0.1
aiormq==6.2.3
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asynctnt==1.2.3
backoff==1.11.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
ConfigArgParse==1.5.3
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.1.3
Faker==12.3.0
fastapi==0.78.0
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
locust==2.8.3
MarkupSafe==2.1.0
msgpack==1.0.3
msgpack-python==0.5.6
multidict==6.0.2
newrelic-telemetry-sdk==0.4.2
opentelemetry-api==1.11.0
opentelemetry-exporter-jaeger==1.11.0
opentelemetry-exporter-jaeger-proto-grpc==1.11.0
opentelemetry-exporter-jaeger-thrift==1.11.0
opentelemetry-exporter-otlp==1.11.0
opentelemetry-exporter-otlp-proto-grpc==1.11.0
opentelemetry-exporter-otlp-proto-http==1.11.0
opentelemetry-ext-newrelic==0.1.1
opentelemetry-instrumentation==0.30b0
opentelemetry-instrumentation-asgi==0.30b0
opentelemetry-instrumentation-fastapi==0.30b0
opentelemetry-instrumentation-requests==0.30b0
opentelemetry-proto==1.11.0
opentelemetry-sdk==1.11.0
opentelemetry-semantic-conventions==0.30b0
opentelemetry-util-http==0.30b0
pamqp==3.1.0
protobuf==3.20.1
psutil==5.9.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==0.9.3
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rfc3986==1.5.0
roundrobin==0.0.2
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tarantool==0.6.5
thrift==0.16.0
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.16.0
websockets==10.3
Werkzeug==2.0.3
wrapt==1.14.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
